# api-docs.yml
openapi: 3.0.0
servers:
  - description: 42 MANITO API DEV SERVER
    url: "https://api.dev.42manito.com/"
  - description: 42 MANITO API SERVER
    url: "https://api.42manito.com/"
info:
  version: "1.0.0"
  title: 42-manito-api
  description: 42 manito api server
paths:
...
  /users:
    get:
      summary: Get all users
      tags:
        - Users
      security:
        - NormalUser: []
      parameters:
        - in: query
          name: take
          description: take `take` Users (`min = 1, max = 100, default = 20`)
          schema:
            type: integer
            format: int32
        - in: query
          name: page
          description: skip `take * page` Users and take `take` Users (`skip >= 0`)
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsersGet'
    post:
      summary: Create a new user
      tags:
        - Users
      security:
        - AdminUser: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersGet'
        '400':
          description: Invalid request body
        '401':
          description: Unauthorized
        '409':
          description: already exists
  /users/{id}:
    get:
      summary: Get a user by ID
      tags:
        - Users
      security:
        - NormalUser: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersGet'
        '400':
          description: Invalid request parameter id
        '404':
          description: User not found
    patch:
      summary: Update user
      tags:
        - Users
      security:
        - OwnerUser: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersGet'
        '400':
          description: Invalid request body
        '401':
          description: Unauthorized
        '409':
          description: already exists
  /users/{id}/reservations:
    get:
      summary: Get a user reservations by ID
      tags:
        - Users
      security:
        - OwnerUser: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - in: query
          name: take
          description: take `take` Users (`min = 1, max = 100, default = 20`)
          schema:
            type: integer
            format: int32
        - in: query
          name: page
          description: skip `take * page` Users and take `take` Users (`skip >= 0`)
          schema:
            type: integer
            format: int32
        - in: query
          name: as_mentor
          description: find `reservations` that reserved as mentor `(default = "true")`
          schema:
            type: string
            format: boolean
        - in: query
          name: as_mentee
          description: find `reservations` that reserved as mentee `(default = "true")`
          schema:
            type: string
            format: boolean
        - in: query
          name: active
          description: find all active (`REQUEST`, `ACCEPT`, `PENDING`) user's `reservations` `(default ="false")`
          schema:
            type: string
            format: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersReservationsGet'
        '400':
          description: Invalid request parameter id
        '404':
          description: User not found
  /users/verify_nickname/{nickname}:
    get:
      summary: Verify nickname
      tags:
        - Users
      parameters:
        - name: nickname
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Invalid nickname
        '409':
          description: Nickname not found
  /mentor_profiles:
    get:
      summary: Get all mentor profiles
      tags:
        - Mentor Profiles
      security:
        - AdminUser: [ ]
      parameters:
        - in: query
          name: take
          description: take `take` Users (min = 1, max = 100, default = 20)
          schema:
            type: integer
            format: int32
        - in: query
          name: page
          description: skip `take * page` Users and take `take` Users (>= 0)
          schema:
            type: integer
            format: int32
        - in: query
          name: is_hide
          description: filter by is_hide
          schema:
            type: boolean
        - in: query
          name: hashtag_id
          description: filter by hashtag
          schema:
            type: integer
            format: int32
        - in: query
          name: category_id
          description: filter by category
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorProfileGet'
  /mentor_profiles/{id}:
    get:
      summary: Get a mentor profile by ID
      tags:
        - Mentor Profiles
      parameters:
        - name: id
          in: path
          required: true
          description: user id
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
        '400':
          description: Invalid request parameter id
        '404':
          description: Mentor profile not found
    patch:
      summary: Update mentor profile
      description: |
        update mentor profiles.
        **NOTE**: **hashtags** and **categories** are replaced with the new ones
      tags:
        - Mentor Profiles
      security:
        - OwnerUser: []
      parameters:
        - name: id
          in: path
          description: user id
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorProfileUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorProfileGet'
        '400':
          description: Invalid request body
        '401':
          description: Unauthorized
        '409':
          description: already exists
...
  /search/mentor/{search_string}:
    get:
      summary: Search mentor by string
      tags:
        - Search
      parameters:
        - in: path
          name: search_string
          required: true
          description: "search mentor profiles that contains search string in nickname or hashtags"
          schema:
            type: string
        - in: query
          name: take
          description: take `take` Users `min = 1, max = 100, default = 20`
          schema:
            type: integer
            format: int32
        - in: query
          name: page
          description: skip `take * page` Users and take `take` Users `take >= 0`
          schema:
            type: integer
            format: int32
        - in: query
          name: search_by_user_nickname
          description: search by nickname `default = true`
          schema:
            type: string
            format: "true | false"
        - in: query
          name: search_by_hashtag_name
          description: search by tag name `default = true`
          schema:
            type: string
            format: "true | false"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchMentorGet'
...
components:
  schemas:
...
    UserCreate:
      type: object
      description: "only for admin"
      properties:
        email:
          type: string
          format: email
          description: "required"
        nickname:
          type: string
          description: "required"
        profileImage:
          type: string
          description: "required"
        role:
          description: "required"
          type: string
          format: enum
          enum:
            - ADMIN
            - USER
    UserUpdate:
      type: object
      properties:
        profileImage:
          type: string
          description: "optional"
    UsersGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
          format: email
        nickname:
          type: string
        profileImage:
          type: string
        role:
          description: "required"
          type: string
          format: enum
          enum:
            - ADMIN
            - USER
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        mentorProfile:
          $ref: '#/components/schemas/MentorProfileSimpleGet'
    UsersReservationsGet:
      type: object
      properties:
        menteeReservations:
          type: array
          items:
            $ref: '#/components/schemas/ReservationGet'
        mentorReservations:
          type: array
          items:
            $ref: '#/components/schemas/ReservationGet'

    ## MentorProfile
...
    MentorProfileGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
        userId:
          type: integer
          format: int32
        shortDescription:
          type: string
        description:
          type: string
        isHide:
          type: boolean
        mentoringCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hashtags:
          type: array
          items:
            $ref: '#/components/schemas/HashtagGet'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryGet'
        user:
          type: object
          properties:
            id:
              type: integer
              format: int32
            nickname:
              type: string
            profileImage:
              type: string
    MentorProfileSimpleGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
        shortDescription:
          type: string
        description:
          type: string
        isHide:
          type: boolean
        mentoringCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hashtags:
          type: array
          items:
            $ref: '#/components/schemas/HashtagGet'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryGet'
    HomeGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
        shortDescription:
          type: string
        description:
          type: string
        isHide:
          type: boolean
        mentoringCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hashtags:
          type: array
          items:
            $ref: '#/components/schemas/HashtagGet'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryGet'
        user:
          $ref: '#/components/schemas/UserSimpleGet'
    UserSimpleGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nickname:
          type: string
        profileImage:
          type: string
    MentorProfileUpdate:
      type: object
      properties:
        isHide:
          type: boolean
          description: "optional, default false"
        shortDescription:
          type: string
          description: 'optional, 0 < len < 50'
        description:
          type: string
          description: 'optional, 0 < len < 1000'
        hashtags:
          type: array
          description: 'optional, number of hashtag <= 5'
          items:
            type: object
            properties:
              id:
                type: integer
                format: int32
        categories:
          type: array
          description: 'optional'
          items:
            type: object
            properties:
              id:
                type: integer
                format: int32
    ## Reservation
...
    ReservationGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: "reservation id"
        mentorId:
          type: integer
          format: int32
        menteeId:
          type: integer
          format: int32
        categoryId:
          type: integer
          format: int32
        requestMessage:
          type: string
        status:
          type: string
          format: enum
          enum:
            - "REQUEST"
            - "ACCEPT"
            - "CANCEL"
            - "PENDING"
            - "COMPLETE"
        mentorFeedback:
          $ref: '#/components/schemas/MentorFeedbackGet'
        menteeFeedback:
          $ref: '#/components/schemas/MenteeFeedbackGet'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hashtags:
          type: array
          items:
            $ref: '#/components/schemas/HashtagGet'
...
    HashtagGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
...
    MenteeFeedbackGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
        menteeId:
          type: integer
          format: int32
        mentorId:
          type: integer
          format: int32
        reservationId:
          type: integer
          format: int32
        rating:
          type: number
          format: float
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MentorFeedbackGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
        menteeId:
          type: integer
          format: int32
        mentorId:
          type: integer
          format: int32
        reservationId:
          type: integer
          format: int32
        rating:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CategoryGet:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string

    ## Search/Mentor
    SearchMentorGet:
      $ref: '#/components/schemas/HomeGet'
  securitySchemes:
    NormalUser:
      type: apiKey
      in: cookie
      name : "Authorization"
      description: "User Role >= USER"
    OwnerUser:
      type: apiKey
      in: cookie
      name : "Authorization"
      description: "User Role >= ResourceOwner"
    AdminUser:
      type: apiKey
      in: cookie
      name : "Authorization"
      description: "User Role >= ADMIN"